change precision from double to float for minimize binary size

+++ src/UMFbx2011.cpp
@@ -411,7 +411,7 @@ protected:
 };
 
 /**
- * convert FbxAMatrix to DoubleListVec
+ * convert FbxAMatrix to FloatListVec
  */
 static void matrix_to_UMMat44d(UMMat44d& dst, const FbxAMatrix& src_matrix)
 {
@@ -425,7 +425,7 @@ static void matrix_to_UMMat44d(UMMat44d& dst, const FbxAMatrix& src_matrix)
 }
 
 /**
- * convert FbxMatrix to DoubleListVec
+ * convert FbxMatrix to FloatListVec
  */
 static void matrix_to_UMMat44d(UMMat44d& dst, const FbxMatrix& src_matrix)
 {
@@ -1003,7 +1003,7 @@ bool UMFbxLoadImpl::assign_blend_shapes(UMObjectPtr object, UMMesh& mesh, FbxNod
 	//						for (int m = 0; m < vertex_count; ++m)
 	//						{
 	//							FbxVector4 v = fbx_shape->GetControlPointAt(m);
-	//							DoubleList vertex;
+	//							FloatList vertex;
 	//							vertex.push_back(v[0]);
 	//							vertex.push_back(v[1]);
 	//							vertex.push_back(v[2]);
@@ -1033,7 +1033,7 @@ bool UMFbxLoadImpl::assign_blend_shapes(UMObjectPtr object, UMMesh& mesh, FbxNod
 	//							for (int m = 0; m < normal_count; ++m)
 	//							{
 	//								FbxVector4 v = fbx_normals->GetAt(m);
-	//								DoubleList normal;
+	//								FloatList normal;
 	//								normal.push_back(v[0]);
 	//								normal.push_back(v[1]);
 	//								normal.push_back(v[2]);
@@ -1083,7 +1083,7 @@ bool UMFbxLoadImpl::assign_normals(
 	{
 		FbxLayerElement::EMappingMode mode = fbx_normals->GetMappingMode();
 		FbxLayerElement::EReferenceMode ref_mode = fbx_normals->GetReferenceMode();
-		umio::DoubleList normal;
+		umio::FloatList normal;
 		normal.resize(3);
 
 		FbxVector4 n;
@@ -1152,7 +1152,7 @@ bool UMFbxLoadImpl::assign_uvs(
 	{
 		FbxLayerElement::EMappingMode mode = fbx_uv->GetMappingMode();
 		FbxLayerElement::EReferenceMode ref_mode = fbx_uv->GetReferenceMode();
-		umio::DoubleList uv;
+		umio::FloatList uv;
 		uv.resize(2);
 
 		FbxVector2 v;
@@ -1223,7 +1223,7 @@ bool UMFbxLoadImpl::assign_vertex_colors(
 	{
 		FbxLayerElement::EMappingMode mode = fbx_vertex_color->GetMappingMode();
 		FbxLayerElement::EReferenceMode ref_mode = fbx_vertex_color->GetReferenceMode();
-		umio::DoubleList vertex_color;
+		umio::FloatList vertex_color;
 		vertex_color.resize(4);
 
 		FbxColor v;
@@ -1314,7 +1314,7 @@ bool UMFbxLoadImpl::assign_mesh(UMObjectPtr object, FbxNode* node)
 		// vertex
 		for (int i = 0; i < control_point_count; ++i)
 		{
-			umio::DoubleList vertex;
+			umio::FloatList vertex;
 			vertex.push_back(control_points[i][0]);
 			vertex.push_back(control_points[i][1]);
 			vertex.push_back(control_points[i][2]);
@@ -2547,7 +2547,7 @@ bool UMFbxSaveImpl::export_shape(FbxShape* fbx_shape, const UMShape& shape)
 //#endif // WITH_PYTHON
 	for (int i = 0; i < vertex_count; ++i)
 	{
-		const DoubleList& vertex = shape.vertex_list().at(i);
+		const FloatList& vertex = shape.vertex_list().at(i);
 		if (static_cast<int>(vertex.size()) < 3) continue;
 		FbxVector4 v(
 			vertex.at(0), 
@@ -2582,7 +2582,7 @@ bool UMFbxSaveImpl::export_shape(FbxShape* fbx_shape, const UMShape& shape)
 	//		
 	//		for (int i = 0; i < normal_count; ++i)
 	//		{
-	//			const DoubleList& normal = shape.normal_list().at(i);
+	//			const FloatList& normal = shape.normal_list().at(i);
 	//			FbxVector4 n(
 	//				normal.at(0), 
 	//				normal.at(1), 
@@ -2722,7 +2722,7 @@ bool UMFbxSaveImpl::export_mesh(FbxScene* scene, UMObjectPtr object)
 				fbx_mesh->InitControlPoints(vertex_count);
 				for (int i = 0; i < vertex_count; ++i)
 				{
-					const DoubleList& vertex = mesh.vertex_list().at(i);
+					const FloatList& vertex = mesh.vertex_list().at(i);
 					if (static_cast<int>(vertex.size()) < 3) continue;
 					FbxVector4 vertex_point(vertex.at(0), vertex.at(1), vertex.at(2), 1.0);
 					fbx_mesh->SetControlPointAt(vertex_point, i);
@@ -2773,7 +2773,7 @@ bool UMFbxSaveImpl::export_mesh(FbxScene* scene, UMObjectPtr object)
 					fbx_normal->SetReferenceMode(FbxLayerElement::eDIRECT);
 					for (int i = 0; i < normal_count; ++i)
 					{
-						const DoubleList& normal = mesh.normal_list().at(i);
+						const FloatList& normal = mesh.normal_list().at(i);
 						if (static_cast<int>(normal.size()) < 3) continue;
 						FbxVector4 normal_point(normal.at(0), normal.at(1), normal.at(2), 1.0);
 						fbx_normal->GetDirectArray().Add(normal_point);
@@ -2792,7 +2792,7 @@ bool UMFbxSaveImpl::export_mesh(FbxScene* scene, UMObjectPtr object)
 					fbx_uv->SetReferenceMode(FbxLayerElement::eINDEX_TO_DIRECT);
 					for (int i = 0; i < uv_count; ++i)
 					{
-						const DoubleList& uv = mesh.uv_list().at(i);
+						const FloatList& uv = mesh.uv_list().at(i);
 						if (static_cast<int>(uv.size()) < 2) continue;
 						FbxVector2 fbx_vertex_uv(uv.at(0), uv.at(1));
 						fbx_uv->GetDirectArray().Add(fbx_vertex_uv);
